CAN IO, Processing: keep processing data without waiting for a request
By waiting for a request, the processing thread does not keep updating
the consistency of the threshold status, meaning that it takes many
requests (as many as the threshold delay) for the status to change.

TODO: the current code keeps the data mutex almost always locked,
potentially starving the sender thread, which needs to lock it.