Processing: limit the time data mutex is locked as much as possible
The module's code has been changed to enclose the critical section, in
which data is updated, is as small as possible.

Also, the configuration of processing bounds was moved to the
'processing_set_mode' function, so that it can be done only once
after changing the mode, simplifying the 'update_data' function.