Processing: implement 'processing_get_data' function
This function hides the complexity of reading the processed data, which
should only be accessed when available and after locking its mutex. By
having an internal function do that, all the previously exposed
variables can now be hidden, and it becomes guaranteed that the data
will only be accessed properly (i.e. it's no longer possible to read it
without the proper measures).

Some code in the CAN IO module had to be adapted.